+++ src/Framework/ContentPacks/Model/CustomBoardData.cs
@@ -1,15 +1,13 @@
 ﻿using Microsoft.Xna.Framework;
-using System;
+using QuestFramework.Framework.Structures;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace QuestFramework.Framework.ContentPacks.Model
 {
     class CustomBoardData
     {
         public string BoardName { get; set; }
+        public BoardType BoardType { get; set; } = BoardType.Quests;
         public string Location { get; set; }
         public Point Tile { get; set; }
         public Dictionary<string, string> UnlockWhen { get; set; }