+++ src/Structures/CustomBoardTrigger.cs
@@ -1,4 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
+using QuestFramework.Framework.Menus;
+using StardewModdingAPI;
 using StardewValley.Menus;
 using System;
 
@@ -10,6 +12,7 @@ namespace QuestFramework.Framework.Structures
         public Point Tile { get; set; }
         public string BoardName { get; set; }
         public bool ShowIndicator { get; set; } = true;
+        public BoardType BoardType { get; set; } = BoardType.Quests;
 
         public Func<bool> unlockConditionFunc;
 
@@ -17,5 +20,11 @@ namespace QuestFramework.Framework.Structures
         {
             return this.unlockConditionFunc == null || this.unlockConditionFunc();
         }
+    }
+
+    public enum BoardType
+    {
+        Quests,
+        SpecialOrders,
     }
 }